Extract Rust ratchet ratio extraction and comparison building
Move within-run Rust/Python ratio extraction and comparison-group validation
into benchmarks/ratchet_ratio_extraction.py, leaving report-value construction
in the ratchet CLI, and extract _build_scenario_comparisons to cut the
complexity of compare_rust_regressions. Lift the ratchet behaviour-test
fixture scaffolding into a colocated support module.
